+++ tools/cli/embraion/__init__.py
@@ -1,3 +1,3 @@
 """EmbrAIon CLI package."""
 
-__version__ = "0.1.0-dev"
+__version__ = "0.2.0.dev0"